Reviewed code for reseting assign grades to include also recovergradesdefault parameter (https://tracker.moodle.org/browse/MDL-41877)

+++ html/moodle2/mod/assign/locallib.php
@@ -609,7 +609,9 @@ class assign {
 
             //XTEC ************ AFEGIT - To fix bug when reseting assign grades (MDL-41877)
             //2013.10.03  @sarjona
-            $DB->delete_records_select('assign_grades', "assignment IN ($assignssql)", $params);
+            if (!$CFG->recovergradesdefault) {
+                $DB->delete_records_select('assign_grades', "assignment IN ($assignssql)", $params);
+            }
             //************ FI
             $DB->delete_records_select('assign_submission', "assignment IN ($assignssql)", $params);
             $status[] = array('component'=>$componentstr,
